Add JMS TCK integration test suite and related support classes (#2)
This commit introduces a Gradle script to download and execute the Jakarta Messaging (JMS) TCK. The following changes were made:

* Import tasks for downloading, verifying, and unzipping the TCK
* Create tasks for patching TCK configuration and tests
* Add support for Qpid JMS Client, ActiveMQ Client, and Artemis Core JMS Client
* Modify build scripts to handle a variety of configurations including JMS versions, client versions, TCK packages, and test cases.

These changes are made as part of efforts to automate running the JMS TCK against the various supported JMS providers and configurations. The Gradle script ensures that the test suite is correctly set up and executed. The support classes provide an abstraction over the differences among the various JMS clients that can be tested.