Add fail safeguards for missing property files
Introduced a fail-fast mechanism to the build script that throws an exception when property files 'jndi.properties' and 'log4j.properties' are missing. This is achieved by adding a 'mustExist' function ensuring these files' existence before copying. Doing so prevents silent failures during the project build when these essential configuration files are missing. This change was spurred by discussions raised in [this StackOverflow thread](https://stackoverflow.com/questions/62889954/copy-task-doesnt-fail-or-log-when-source-file-is-missing).