Update authentication credentials for broker setup
Changed the authentication credentials for Artemis Broker setup in the Gradle workflow file along with code commentary changes in the broker.kt file. These changes were required to maintain consistency in broker login credentials across all settings and to provide accurate instructions in the README about the broker setup. User and password has been updated to 'tckuser' and references to these changes have been included in the README for guidance. Also, outdated or redundant tasks in the TODO section of the README file have been managed for clarity.